// core/types.ts
export type ThothNode = Node & {
  inspector: Inspector
  display: (content: string) => void
  outputs: { name: string;[key: string]: unknown }[]
  category?: string
  deprecated?: boolean
  displayName?: string
  info: string
  subscription: Function
  console: ThothConsole
}
// ...
export type NodeData = ReteNodeData & {
  fewshot?: string
  display: Function
  error?: boolean
  console: ThothConsole
}

// export type Node = {
//   id: number,
//   data: NodeData,
//   name: string,
//   inputs: NodeOutputs,
//   outputs?: NodeOutputs,
//   position: number[]
// }

export type Spell = {
  id: string
  nodes: Record<number, Node>
}

export type ThothReteInput = {
  type: TaskOutputTypes
  outputData: unknown
  task: ThothTask
  key: string
}

export type TaskOutput = {
  type: TaskOutputTypes
  task: ThothTask
  key: string
}

export type ModuleWorkerOutput = WorkerOutputs & {
  [key: string]: string | string[]
}
export type ThothWorkerInput = string | unknown | unknown[]
export type ThothWorkerInputs = { [key: string]: ThothWorkerInput[] }
export type ThothWorkerOutputs = WorkerOutputs & {
  [key: string]: TaskOutput
}